Move Display method to a separate file
